modification du nuage pour le scale et la disparition progressive maintenant le nuage pleut progressivement

+++ Unity Project/Assets/Script/Planet.cs
@@ -162,7 +162,7 @@ public class Planet : MonoBehaviour
         m_money -= p_cost;
     }
 
-    public void rain(int p_rainingQuantity)
+    public void rain(double p_rainingQuantity)
     {
         m_water += p_rainingQuantity;
     }
